+++ .storybook/preview.ts
@@ -1,5 +1,5 @@
 import type { Preview } from '@storybook/react';
-import '../src/app/styles/globals.css';
+import './storybook.css';
 
 const preview: Preview = {
 	parameters: {
@@ -12,6 +12,9 @@ const preview: Preview = {
 			options: {
 				storySort: (a, b) =>
 					a[1].kind === b[1].kind ? 0 : a[1].id.localeCompare(b[1].id, undefined, { numeric: true }),
+			},
+			nextjs: {
+				appDirectory: true,
 			},
 		},
 	},
