+++ src/components/ButtonEx/storybook/index.tsx
@@ -27,7 +27,7 @@ interface ButtonProps {
 /**
  * Primary UI component for user interaction
  */
-export const Button = ({ primary = false, size = 'medium', backgroundColor, label, ...props }: ButtonProps) => {
+export const ButtonEx = ({ primary = false, size = 'medium', backgroundColor, label, ...props }: ButtonProps) => {
 	const mode = primary ? 'storybook-button--primary' : 'storybook-button--secondary';
 	return (
 		<button type="button" className={['storybook-button', `storybook-button--${size}`, mode].join(' ')} {...props}>
